Nothing really, just trying to figure out how to do this

+++ ParticleToy/CircularWireConstraint.cpp
@@ -3,12 +3,12 @@
 
 #define PI 3.1415926535897932384626433832795
 
-static void draw_circle(const Vec2f &vect, float radius) {
+static void draw_circle(const Vec2f &vect, double radius) {
     glBegin(GL_LINE_LOOP);
     glColor3f(0.0, 1.0, 0.0);
     for (int i = 0; i < 360; i = i + 18) {
-        float degInRad = i * PI / 180;
-        glVertex2f(vect[0] + cos(degInRad) * radius, vect[1] + sin(degInRad) * radius);
+        double degInRad = i * PI / 180;
+        glVertex2f((GLfloat) (vect[0] + cos(degInRad) * radius), (GLfloat) (vect[1] + sin(degInRad) * radius));
     }
     glEnd();
 }
